Complete!

+++ src/scheduling/scan.rs
@@ -7,7 +7,7 @@ pub struct SCAN {
     current: u16,
     max_cylinder: u16,
     direction: Direction,
-    movements: u16,
+    movements: u64,
     reversals: u16,
     circular: bool,
 }
@@ -67,10 +67,11 @@ impl Scheduler for SCAN {
     fn next_request(&mut self) -> Option<Request> {
         let mut min_dist = u16::MAX;
         let mut index: Option<usize> = None;
-        // Find next closest request in the direction of movement
+
         for (i, request) in self.requests.iter().enumerate() {
             let dist = u16::abs_diff(request.location, self.current);
 
+            // Find next closest request in the direction of movement
             match self.direction {
                 Direction::HIGH => {
                     if request.location > self.current && dist < min_dist {
@@ -88,45 +89,38 @@ impl Scheduler for SCAN {
                 },
                 Direction::DEFAULT => panic!("SCAN needs direction")
             }
-            // if dist < min_dist {
-            //     min_dist = dist;
-            //     index = Some(i);
-            // }
         }
 
 
-            if self.movements == 0 {
-                self.movements = 1;
-            }
+        // Count the first movement done
+        if self.movements == 0 {
+            self.movements = 1;
+        }
 
         if !index.is_none() {
 
             let request = &self.requests[index.unwrap()];
-            self.movements += min_dist;
-            // println!("()Next: {}", request.location);
+
+            self.movements += min_dist as u64;
             self.current = request.location;
 
+            // Service the request
             Some(self.requests.remove(index.unwrap()))
         } else {
-            // println!("Done with direction: {}", self.direction);
+
+            // If no more requests were found in the current direction, go to pole and turn around
             match self.direction {
-                // Set direction to be low
                 Direction::HIGH => {
-                // println!("Current: {}", self.current);
-                // println!("Next: {}", self.max_cylinder);
-                // println!("Dist: {}", u16::abs_diff(self.current, self.max_cylinder));
                     self.direction = Direction::LOW;
-                    self.movements += u16::abs_diff(self.current, self.max_cylinder);
+                    self.movements += u16::abs_diff(self.current, self.max_cylinder) as u64;
                     self.reversals += 1;
-                    // TODO: add movements to edge
                     self.current = self.max_cylinder;
                 },
                 Direction::LOW => {
                     self.direction = Direction::HIGH;
-                    self.movements += u16::abs_diff(self.current, 0);
+                    self.movements += u16::abs_diff(self.current, 0) as u64;
                     self.reversals += 1;
                     self.current = 0;
-                    // TODO: add movements to edge
                 },
                 Direction::DEFAULT => panic!("SCAN needs direction")
             }
